refactor(condo): DOMA-7185 refactored names and placement of tasks for "organization" domain
