+++ apps/condo/domains/organization/tasks/index.js
@@ -1,9 +1,18 @@
-const refreshSbbolClientSecret = require('./refreshSbbolClientSecret')
-const { syncSbbolBankAccountsCron: syncSbbolBankAccounts } = require('./syncSbbolBankAccounts')
-const { syncSbbolTransactionsCron: syncSbbolTransactions } = require('./syncSbbolTransactions')
+const { createCronTask, removeCronTask, createTask } = require('@open-condo/keystone/tasks')
+
+const { refreshSbbolClientSecret } = require('./refreshSbbolClientSecret')
+const { syncSbbolBankAccounts } = require('./syncSbbolBankAccounts')
+const { syncSbbolTransactions } = require('./syncSbbolTransactions')
+const { syncSbbolTransactionsForToday } = require('./syncSbbolTransactionsForToday')
+
+// Renamed to "syncSbbolTransactionsForToday"
+removeCronTask('syncSbbolTransactionsCron', '0 0 * * *')
+// Renamed to "syncSbbolBankAccounts"
+removeCronTask('syncSbbolBankAccountsCron', '0 0 * * *')
 
 module.exports = {
-    syncSbbolTransactions,
-    refreshSbbolClientSecret,
-    syncSbbolBankAccounts,
+    refreshSbbolClientSecretCronTask: createCronTask('refreshSbbolClientSecret', '0 1 * * *', refreshSbbolClientSecret),
+    syncSbbolBankAccountsCronTask: createCronTask('syncSbbolBankAccounts', '0 0 * * *', syncSbbolBankAccounts, { priority: 2 }),
+    syncSbbolTransactionsForTodayCronTask: createCronTask('syncSbbolTransactionsForToday', '0 0 * * *', syncSbbolTransactionsForToday),
+    syncSbbolTransactionsTask: createTask('syncSbbolTransactions', syncSbbolTransactions, { priority: 2 }),
 }
