+++ apps/condo/domains/organization/tasks/syncSbbolTransactions.js
@@ -1,18 +1,12 @@
 const dayjs = require('dayjs')
-const get = require('lodash/get')
-const isEmpty = require('lodash/isEmpty')
+const { get } = require('lodash')
 
-const { getLogger } = require('@open-condo/keystone/logging')
 const { getSchemaCtx } = require('@open-condo/keystone/schema')
-const { createTask, createCronTask } = require('@open-condo/keystone/tasks')
 
 const { BANK_SYNC_TASK_STATUS } = require('@condo/domains/banking/constants')
 const { BankSyncTask } = require('@condo/domains/banking/utils/serverSchema')
-const { SBBOL_IMPORT_NAME } = require('@condo/domains/organization/integrations/sbbol/constants')
 const { requestTransactions } = require('@condo/domains/organization/integrations/sbbol/sync/requestTransactions')
-const { OrganizationEmployee } = require('@condo/domains/organization/utils/serverSchema')
-const { SBBOL_IDP_TYPE } = require('@condo/domains/user/constants/common')
-const { UserExternalIdentity } = require('@condo/domains/user/utils/serverSchema')
+
 
 const dvAndSender = { dv: 1, sender: { dv: 1, fingerprint: 'syncSbbolTransactions' } }
 
@@ -23,54 +17,15 @@ async function updateStatusOfBankSyncTask (context, taskId, status) {
     })
 }
 
-const logger = getLogger('sbbol/CronTaskSyncTransactions')
-
 /**
  * Synchronizes SBBOL transaction data with data in the system
  */
-async function syncSbbolTransactions (dateInterval) {
+async function syncSbbolTransactions (taskId) {
     const { keystone: context } = await getSchemaCtx('User')
-    // TODO(VKislov): DOMA-5239 Should not receive deleted instances with admin context
-    const usersWithSBBOLExternalIdentity = await UserExternalIdentity.getAll(context, {
-        identityType: SBBOL_IDP_TYPE,
-        deletedAt: null,
-    })
-    if (isEmpty(usersWithSBBOLExternalIdentity)) return logger.info('No users imported from SBBOL found. Cancel sync transactions')
-
-    const syncedOrgIds = []
-    for (const identity of usersWithSBBOLExternalIdentity) {
-        const userId = identity.user.id
-        const [employee] = await OrganizationEmployee.getAll(context, {
-            user: { id: userId },
-            organization: {
-                id_not_in: syncedOrgIds,
-                importRemoteSystem: SBBOL_IMPORT_NAME,
-                deletedAt: null,
-            },
-            deletedAt: null,
-        }, { first: 1 })
-
-        if (employee) {
-            const organization = get(employee, 'organization')
-            await requestTransactions({
-                dateInterval,
-                userId,
-                organization,
-            })
-            syncedOrgIds.push(organization.id)
-        }
-    }
-}
-
-/**
- * Synchronizes SBBOL transaction data with data in the system
- */
-async function syncSbbolTransactionsBankSyncTask (taskId) {
     if (!taskId) {
         await updateStatusOfBankSyncTask(context, taskId, BANK_SYNC_TASK_STATUS.ERROR)
         throw new Error('Missing taskId')
     }
-    const { keystone: context } = await getSchemaCtx('User')
 
     let bankSyncTask
 
@@ -98,9 +53,5 @@ async function syncSbbolTransactionsBankSyncTask (taskId) {
 }
 
 module.exports = {
-    syncSbbolTransactionsCron: createCronTask('syncSbbolTransactionsCron', '0 0 * * *', async () => {
-        const date = dayjs().format('YYYY-MM-DD')
-        await syncSbbolTransactions([date])
-    }),
-    syncSbbolTransactionsBankSyncTask: createTask('syncSbbolTransactionsBankSyncTask', syncSbbolTransactionsBankSyncTask, { priority: 2 }),
+    syncSbbolTransactions,
 }
