+++ apps/condo/domains/organization/tasks/syncSbbolTransactionsForToday.js
@@ -1,60 +1,61 @@
-const get = require('lodash/get')
-const isEmpty = require('lodash/isEmpty')
+const dayjs = require('dayjs')
+const { get, isEmpty } = require('lodash')
 
 const { getLogger } = require('@open-condo/keystone/logging')
 const { getSchemaCtx } = require('@open-condo/keystone/schema')
 
-const { BANK_INTEGRATION_IDS } = require('@condo/domains/banking/constants')
-const { BankIntegration } = require('@condo/domains/banking/utils/serverSchema')
 const { SBBOL_IMPORT_NAME } = require('@condo/domains/organization/integrations/sbbol/constants')
-const { syncBankAccounts } = require('@condo/domains/organization/integrations/sbbol/sync/syncBankAccounts')
+const { requestTransactions } = require('@condo/domains/organization/integrations/sbbol/sync/requestTransactions')
 const { OrganizationEmployee } = require('@condo/domains/organization/utils/serverSchema')
 const { SBBOL_IDP_TYPE } = require('@condo/domains/user/constants/common')
 const { UserExternalIdentity } = require('@condo/domains/user/utils/serverSchema')
 
 
-const logger = getLogger('sbbol/syncBankAccounts')
+const logger = getLogger('syncSbbolTransactionsForToday')
 
 /**
- * Syncs bank accounts with SBBOL accounts
+ * Synchronizes SBBOL transaction data with data in the system
  */
-async function syncSbbolBankAccounts () {
+async function syncSbbolTransactionsForDateInterval (dateInterval) {
     const { keystone: context } = await getSchemaCtx('User')
     // TODO(VKislov): DOMA-5239 Should not receive deleted instances with admin context
     const usersWithSBBOLExternalIdentity = await UserExternalIdentity.getAll(context, {
         identityType: SBBOL_IDP_TYPE,
         deletedAt: null,
     })
-    if (isEmpty(usersWithSBBOLExternalIdentity)) return logger.info('No users imported from SBBOL found. Cancel sync bank accounts')
-
-    const integration = await BankIntegration.getOne(context, { id: BANK_INTEGRATION_IDS.SBBOL })
-    if (!integration) throw new Error(`Cannot find SBBOL integration by id=" ${BANK_INTEGRATION_IDS.SBBOL}"`)
+    if (isEmpty(usersWithSBBOLExternalIdentity)) return logger.info('No users imported from SBBOL found. Cancel sync transactions')
 
     const syncedOrgIds = []
-    for (let identity of usersWithSBBOLExternalIdentity) {
+    for (const identity of usersWithSBBOLExternalIdentity) {
+        const userId = identity.user.id
         const [employee] = await OrganizationEmployee.getAll(context, {
-            user: {
-                id: identity.user.id,
-            },
+            user: { id: userId },
             organization: {
                 id_not_in: syncedOrgIds,
                 importRemoteSystem: SBBOL_IMPORT_NAME,
                 deletedAt: null,
             },
             deletedAt: null,
-            isRejected: false,
-            isBlocked: false,
         }, { first: 1 })
 
         if (employee) {
             const organization = get(employee, 'organization')
-            await syncBankAccounts(identity.user.id, organization)
+            await requestTransactions({
+                dateInterval,
+                userId,
+                organization,
+            })
             syncedOrgIds.push(organization.id)
         }
     }
-
 }
 
+async function syncSbbolTransactionsForToday () {
+    const date = dayjs().format('YYYY-MM-DD')
+    await syncSbbolTransactionsForDateInterval([date])
+}
+
+
 module.exports = {
-    syncSbbolBankAccounts,
+    syncSbbolTransactionsForToday,
 }
